Moved ThreadBox into core.internal
