+++ node/src/main/kotlin/net/corda/node/services/events/NodeSchedulerService.kt
@@ -1,7 +1,7 @@
 package net.corda.node.services.events
 
 import com.google.common.util.concurrent.SettableFuture
-import net.corda.core.ThreadBox
+import net.corda.core.internal.ThreadBox
 import net.corda.core.contracts.SchedulableState
 import net.corda.core.contracts.ScheduledActivity
 import net.corda.core.contracts.ScheduledStateRef
